feat: merge system and backoffice
